requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
werkzeug
openai>=1.0.0

# Optional utilities
json-repair

